// web/src/api/generated/types.gen.ts
export type ComponentProperties = {
    /**
     * Chemicalformula
     *
     * Chemical name of the component
     */
    chemicalFormula: string;
    /**
     * Altname
     *
     * Alternate name of the component
     */
    altName: string;
    /**
     * Molecularweight
     *
     * Molecular weight of the component (g/mol)
     */
    molecularWeight: number;
};

/**
 * ComponentResponse
 *
 * Model for containing the response of getting the component dictionary.
 */
export type ComponentResponse = {
    /**
     * Components
     *
     * Dictionary of component_ids as string and ComponentProperties as value
     */
    components: {
        [key: string]: ComponentProperties;
    };
};